Be more consistent about icons in templates
The templates behaved a bit odd with replaced spells, where the icon
depended on whethera replacement (that is talent) was active on loading
the database.

The best solution would probably be to always display the base icon, but
that seems to be impossible. Instead update the icon to always show the
correct icon for the current state.

Github-Issue: 570