Tweak FrameChooser to choose child auras over groups
If multiple auras match are under the mouse cursor, this will prefer
normal auras to groups.

We use the FrameChooser for attaching and action/glow. For the later
choosing childs is probably 90% correct. For the later, attaching to a
group should be rare, because a child is "attached" to its group by
being a member.

Noticed in the comments of
Github-Issue: 572