Add a "exact match" for spell ids option to a few triggers
Both Guardian and Feral druids have a "Trash" spell with a cooldown,
and with the right talents, druids have access to both spells.

Thus they want to track the cooldown of the right "Trash".

So add some code to enable that as a option and rewrite a few triggers
to offer that option, namely: Action Usable, Cooldown Progress (Spell),
Cooldown Ready and Charges Changed.

Technically the Cooldown Ready and Charges Changed spell change
behaviour with this patch, as they used to only check via the spellid
previously. But this is more a bug fix, similary to what was done
for Cooldown Progress previously.

Github-Issue: 630