Don't schedule a CooldownScan if the remaingtime can't decrease anymore
The aura in the ticket has the remaining time option checked, without
a time. Since thus the remainingTime would be 0 and the remainingCheck
would be 0, this would lead to excessive ScheduleCooldownScan calls.

If the remaingtime is already 0 (or negative), it won't decrease
anymore, so scheduling any checks is pointless.

Github-Issue: 710