Fix incorrect conditions deactivation for fallback states
In 80f821f4824 the way deactivate conditions are handled was changed.

That surfaced a problem in how fallback states were treated.

The case in which the bug surfaced is:
A aura should be shown, because over all the trigger activation is true,
but the trigger that provides dynamic information is not active.

In those cases a fallback state is created and used. This fallback state
is created in ApplyStatesToRegion.

By not running conditions on all regions, but only on the ones which are
hiding, the correct condition evaluation is done in ApplyStatesToRegions

Github-Issue: 716