Fix Action Usable for unknown spells
In a30f2fb70d12e59840960c5eadb9a805d45b4def I added a spell known check
to Action Usable and Cooldown Progress. With that patch any triggers for
unknown spells would always return false.

But users were using inverse Action Usable to check if they had not
selected a talent. So make Action Usable inverse trigger if the spell
is unknown.

This patch does not touch Cooldown Progress, because technically the
spell is not on cooldown when it is unknown.

Ticket-Nr: 1229