Fix group triggers unloading
Group triggers are stored in separate structures, so we need to remove
it from there too on unloading.

The way group triggers are slightly different from normal triggers for
match functions. This is unfortunate, but requires a bit of refactoring to
fix, and that's not something that should be done before the first
release.

Github-Issue: #893