Use KuiNameplatesPlayerAnchor for kui nameplates
As that fixes the a variant of text vanishing bug described in the
linked bug report:

A icon attached to the PRD would loose its cooldown text due to zooming,
due to how positioning of nameplates in kuis work. The actual
KuiNameplatesPlayerAnchor is positioned differently and thus doesn't
have that problem.

This might change the positions for auras attached to the PRD for kui
users, but is probably not too much of a problem.

Github-Issue: #648