Change how cloneIds for cloning are dertermined
The issue in the linked ticket is that states cloneId are just a number
that gets incremented for each match. Thus if we go from 2 matches to 1
match, the second state gets removed. This can be either match.

Improve that, by using se unit+spellId as the cloneId, so that auras
keep reusing their states even if their index changes,
but also different auras don't reuse states.

The same problem still remains for cases where a unit can have multiple
buffs/debuffs with the same spellId, but that should be rare and there's
not much we can do to distinguish them.

Github-Issue: #974