Item Input: Don't validate numbers at all
We used GetItemInfo to validate the input, but due to the "recent"
changes making GetItemInfo async, this doesn't work too well.

So for item ids, just assume that they are valid items. This allows
users to always enter items via their id.

The right fix would be to somehow wait for the GET_ITEM_INFO_RECEIVED
event and then validate. That sounds impossible to do though.

Ticket-Nr: 1275