Fix Shield of the Righteous cooldown tracking
Some spells like SotR that are not on the gcd have a small ~1s
cooldown to prevent the user from accidentally triggering it multiple
times.

Since 6160b7b we prefer showing the GetSpellCooldown instead of
GetSpellCharges, but that isn't useful for these minor cooldowns.

Thus prefer to show GetSpellCharges if the cooldown is less than a
second long.

Ticket-Nr: 1303