Optimize and rewrite dynamic groups
This is a rewrite of the dynamic group code (in particular the sorting and positioning code), in order to do less work.
In particular, dynamic groups are less eager to completely reset all of their data.
The sorting algorithm is a differential variant of insertion sort.

With these changes, [the most popular set of raiding auras](https://wago.io/bfaraid2)
would consume an estimated 2-5 fewer seconds of CPU time during an average raid encounter.
What's more, since this optimization reduces how much is done each update and not their frequency,
this should reduce the amount of stuttering which dynamic groups can cause
which is the most visible kind of performance problem.