fix user mode value retrieval
We were incorrectly just checking config[references] for merged stuff.
But, config[references] is a table with keys, and we might not have merged all of the keys.
So, base the dereferencing on option[references], which is how we build the merged user mode anyways.
This has some knock-on effects though, so some more helper methods had to be written.

Fixes #1160