Fix showing/enabling of various precions/custom text options
A few places used a simple find to check for %p or %t. As the ticket
points out %{p} is nowadays also a valid syntax.

Also a few places used WeakAuras.ContainsPlaceHolders with "c", but
that's incorrect since those options should also be shown for e.g. %c2

And a few icon options didn't check that the texts were actually
enabled.

Fixes: #1196