Make cooldown tracking work like in 2.11.3
If we update a spell cooldown while it's cooldown is shorter than the
gcd, the blizzard api will return the gcd cooldown.

Thus it's impossible to know whether the cooldown is truely reset, or
just hidden.

With this patch and in 2.11.3, we just assume that the cooldown reset.
Whereas before this patch, we would continue the cd.

There isn't really a right answer for this.

Fixes: #1218