enhance parent-child validation
Once upon a time, there was an addon called SpeedrunSplits.
It was a good addon, used by many, in oldschool WoW.
Then, Blizzard decided to make oldschool WoW new again, and a lot of people were very happy.
The author of SpeedrunSplits decided to make his addon available for Classic WoW, and again, a lot of people were very happy.
But then, it was discovered that SpeedrunSplits broke the WeakAuras database accidentally, and a lot of people were very sad.
So, we the developers of WeakAuras will make the database more robust against what broke,
And hopefully people will be happy again.

...More seriously, the bug here was that SpeedrunSplits was replacing the tContains function with another
which had incompatible behavior, thus confusing SyncParentChildRelationships into thinking that groups
were corrupted. This issue has been fixed in SpeedrunSplits beta 2.4 and later, but the damage to the database remains.
So, improve SyncParentChildRelationships to detect the damage and repair it.
Also incidentally remove the vulnerability in the first place.