Model for AuraBar sub element
* Can be in front or behind the foreground texture
* Can be clipped by the progress

There are no options for SetTransform/SetPosition, which limits which
models are usable, but the apis are strange, so that's for another
patch.

The ModelPicker changes are not ideal, the ModelPicker is a bit to
married to the Model. Though both issues can be solved if there's a need
for finer control.

Fixes: #407