Fix Action Button Glow to look like 2.15.5
Depending on the glow type, we actually want to use a base to attach the
glow. For the Action Button Glow we use the whole region as we did for
all glows in 2.15.5 and until bbd9a7a697.

That had the issue that the pixel and auto cast shine glows would be
misplaced with masque skins, as they need to use the icon and not the
whole region as the anchor. So we keep that behaviour for the other two
glows.

Fixes: #1872