Fix MoverSizer breakage
Updating the states while the MoverSizer is resizing breaks the
MoverSizer. Because we hide the fake states which hides the region,
which aborts the resizing. Fix this by making the WeakAuras.Add from the
MoverSizer special. They don't need to do any of the state updating as
these cannot have changed from the MoverSízer

Fixes: #1884