Fix zone, zoneGroup and ecounter tooltip in load options
Apparently the API is still stateful and tied to whatever is open in the
encounter journal. E.g. if the dungeon tab is open EJ_GetEncounterInfoByIndex
does not return any data for raids.

Circumvent that, by simply iterating the encounters/dungeonds before the
encounter jounal is loaded. That is, just after login.

Fixes: #1930