Fix Masque integration with clones and zoom/keep aspect ratio
We fixup the texcoords that masque sets if zoom or aspect ratio is
enabled. (WHich is a bit questionable, but well it's there.)

We used to call ReSkin on the msq group, which would change the skin for
all clones and thus remove our fixup. (And would be qudratic)

Instead just tell masque to reskin one button, so that we can then fix
it.

Fixes: #1969