Fix dynamic group limit option
This was broken by the introduction of the per frame anchor code.

Previously the iteration was over all children and for each child check
the position returned by the grow function. A missing position, thus
meant that the child should be hidden.

The iteration in DoPositionChildrenPerFrame is now over the positions
for a specific frame and thus a missing child is never hidden.

The easiest way to fix it, is to change the interface between grow functions and
dynamic group slightly. The grow function should now explicitly assign a
falsey value to show for children beyond the limit.

Since this is only a edge case, that no one has found in months the api
change is probably the best course

Fixes: #2024