Revert "Don't check for state.show in ReplacePlaceHolders"
Since this breaks text replacements,

If a state is set to show = false, it needs to be ignored for text
replacements, as otherwise stale information could be shown.

But, there's one important exemption. For auras that send a chat message
in their OnHide code, we still want to use the state information.

So, introduce a way for the ReplacePlaceHolders to know about this
special case. This is a bit hacky but unfortunately what users would
expect.

Fixes: #2186

This reverts commit 13c2eaced88bd8fe244b1a2285cabc79a9bd450f.