Add advanced Name/Realm filtering to Health/Power/Cast/Charcateristics
That is allow multiple matching names/realms similar to the load option.

Don't offer blacklisting names/realms for now, as that seems more niche
to me and the only reason load has it, is because the old system allowed
it.

This breaks backwards compability in a minor way.

The old conditions allowed for checking if the unit name matched a lua
pattern, or via a contains check. That's no longer supported, instead
it's always a full name check.

Since it was only supported via conditions and not in the trigger
settings, I consider this more of a bug fix.