Be less brain dead for updating the "eye" texture in the Options
For reasons lost to time and likely decades, the eye textures
open/close/half-open texture was set in a OnUpdate handler.

So every frame.
For all auras.
And for groups by iterating over all children.

Fix this by updating them only if needed. Athough this can be optimized
further and made nicer, this fixes the memory usage observed in the
ticker already.

Fixes: #3161