Make spellcache take up less constants
Unfortunately our spell cache leads to such a enourmous options saved
variables, that on next interface load the data cannot be loaded,
leading to a "constant table overflow".

The root cause of that error is the amount of unique constants used in
the saved variables, which is limited to 2^18.

Fix that by keeping the final layer of the table in a string, instead of
a table. This severly reduces the amount of unique constants at some run
time cost.

Fixes: #3503