Fix corner case in anchoring auras
The game used to detect cycles in anchoring frames, but apparently it no
longer does.

In this case a group was anchored to own of its own children, thus
creating a cycle.

Fix that by generally disallowing anchoring to the group's children.

Doing so does not always result in a cycle, but we disallow it anyway.

It's also possible to create cycles without using parent/child
relationships, which we don't detect here.

Fixes: #3971
Fixes: #3957