Texture: Enable rotation/mirroring for atlas textures
This "properly" fixes atlas textures to work as good as they can.

We want to allow rotation/mirroring for atlas textures, but only by 90°
angles, since we allowed that in the past. So essentially for atlas
textures, we pretend that data.rotate is false.

Technically this breaks when a state changes the texture type (atlas or
not atlas) via the state. But that's an edge case that we can figure out
if its reported...

Fixes: #4101