Fix customTextUpdate not being run in some circumstances
The custom text functions are shared between all text sub elements.
Previously the first sub element was responsible for running the custom
text function.

Since text sub elements now unregister their handlers if they aren't
visible we need to be smarter.

But the "smarter" way of checking the last update time is not smart
enough because multiple events can happen at the same GetTime()
timepoint. In that case we need to rerun the custom text function.

So make the regionPrototype clear the last updated time point, this
ensures the the first text sub element will run the custom text
function.

Having the main region know about the shared custom text function is
annoying, but for now the best I can come up with.

Fixes: #4261