Change EnsureRegion to not create child regions for dynamic groups
We need to ensure that aura data is first Added() before SetRegion() is
called. For reasons that I don't remember EnsureRegion for a dynamic
group would call EnsureRegion on all children.

Since we do a top-down Adding of auras, this would result in trying to
create regions for auras that weren't yet added.

Thus remove that and see what breaks...

Fixes: #4928