Fix lua error for OnHide Animations in certain cases
By reverting the recent change in 15999aa03bd902, which introduced a
C_Timer.

Instead move things around until they run in the right order. In other
words this needs cleaning up with a proper refactoring.

The right oder is:
* Expand code -> Dynamic group border calculation
* OnUpdate of BT2
* OnUpdate of Animations

The OnUpdate code of BT2 needs to run before Animations, so that auras
that have a fade in animation get the fade in animation applied in the
first frame.

The Dynamic group border calculation needs to run before the Animations
OnUpdate, because it takes the current position.

And the OnUpdate code is naturally later than code that runs on other
events.

Fixes: #4965