improve AddMany data flow
otherwise, there's an edge case if Modernize is broken:
1. WeakAuras notices that WeakAurasSaved.dbVersion < internalVersion
2. AddMany is invoked with takeSnapshots == true, and thus calls SetMigrationSnapshot
3. Archivist successfully creates an archive, and because storeType == "ReadOnly",
   the aura data becomes an Archivist store object per spec
4. AddMany calls PreAdd, which notices data.internalVersion < internalVersion,
   & calls Modernize
5. Modernize, being broken, errors out & never updates data.internalVersion
6. AddMany calls Add, which notices data.internalVersion < internalVersion,
   & calls SetMigrationSnapshot a second time (preparing to Modernize)
7. Archivist notices that the object passed to it during CreateStore
   is already a store object, and barfs.

really, this error is the correct behavior of Archivist; it revealed a flaw
in how WeakAuras manages the upgrading of the database. the correct behavior
should be that Modernize is called precisely once. If successful,
then we move on with normal processing. If it fails, then we should not
attempt to Modernize again, and we *especially* shouldn't be setting the
snapshot a second time with potentially corrupted data. that would defeat
its whole purpose!

essentially, the problem is that while PreAdd is meant to be idempotent,
Add is not. so, we make two changes:
1. Invoke pAdd directly from AddMany, instead of calling PreAdd & then Add.
2. use the same bad aura detection logic in the loop that calls pAdd as
   in the one that calls PreAdd.