Fix Spell Cooldowns override tracking
The determination of an spell overriding another spell was done wrongly
in WatchSpellCooldown
Thus for the spell "Crusade" the override was determined to be 231895.
This lead to both Crusade and 231895 being watched, but depending on the
order in which CheckSpellCooldown checked those spells, WeakAuras.GetSpellCooldown
could return the wrong data.

Fixes: #5173